2025/main_page/requirements.txt
pip
wheel
setuptools
selenium
slackclient
pillow
